refactor: 중복 사용되는 타입 리팩터링
