+++ hooks/src/lib/useCVC.ts
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import useInput from './useInput';
 import { validateCVCFormat, validateNumber } from './validator';
-import { UseCard } from './types';
+import { UseCard } from './type';
 
 const useCVC = (initialValue: string): UseCard => {
   const { value, setValue, handleChange, handleBlur, errorInfo } = useInput(initialValue, {
