+++ hooks/src/lib/useExpiryDate.ts
@@ -1,16 +1,13 @@
+import { Options } from './type';
 import useExpiryMonth from './useExpiryMonth';
 import useExpiryYear from './useExpiryYear';
 
-interface Options {
-  month?: {
-    isAutoFocus?: boolean;
-  };
-  year?: {
-    isAutoFocus?: boolean;
-  };
+interface OptionProps {
+  month?: Options;
+  year?: Options;
 }
 
-const useExpiryDate = (initialValue: { month: string; year: string }, options?: Options) => {
+const useExpiryDate = (initialValue: { month: string; year: string }, options?: OptionProps) => {
   const month = useExpiryMonth(initialValue.month, { ...options?.month });
   const year = useExpiryYear(initialValue.year, { ...options?.year });
