+++ hooks/src/lib/useExpiryMonth.ts
@@ -1,11 +1,7 @@
 import useInput from './useInput';
 import { useEffect } from 'react';
 import { validateExpiryMonthFormat, validateNumber } from './validator';
-import { UseCard } from './types';
-
-interface Options {
-  isAutoFocus?: boolean;
-}
+import { Options, UseCard } from './type';
 
 const useExpiryMonth = (initialValue: string, options?: Options): UseCard => {
   const { value, setValue, handleBlur, errorInfo, setErrorInfo } = useInput(initialValue, {
