+++ hooks/src/lib/useExpiryYear.ts
@@ -1,11 +1,7 @@
 import useInput from './useInput';
 import { useEffect } from 'react';
 import { validateExpiryYearFormat, validateNumber } from './validator';
-import { UseCard } from './types';
-
-interface Options {
-  isAutoFocus?: boolean;
-}
+import { Options, UseCard } from './type';
 
 const useExpiryYear = (initialValue: string, options?: Options): UseCard => {
   const { value, setValue, handleBlur, errorInfo, setErrorInfo } = useInput(initialValue, {
